requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-SSLify==0.1.5
Flask-Session>=0.5.0
bcrypt==4.2.1
bleach==6.1.0
SQLAlchemy==2.0.25
requests==2.32.0
Werkzeug==3.0.6
PyGithub==1.59.0
python-dotenv==1.0.0